+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2023-10-25
+### Fixed
+ - Blocking of application in case data is available [#7]
+
 ## [0.2.0] - 2023-06-09
 ### Fixed
 - Changed type of BMS temperature to int8 [#3]